requirements.txt
matplotlib==3.8.3
numpy==1.26.4
opencv_contrib_python==4.9.0.80
opencv_contrib_python_headless==4.9.0.80
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
python_pptx==0.6.23
torch==2.2.1+cu121
tqdm==4.66.2
TTS==0.22.0
